Bot methods are now another file to make adding new ones faster
